set up storage for blocks

+++ backend/crates/storage_postgres/src/query_services/block_query_service.rs
@@ -21,7 +21,7 @@ impl From<OpenedBlockModel> for OpenedBlockDto {
             id: model.id,
             title: model.title,
             opened_at: model.opened_at,
-            tab_index: model.tab_index as u32,
+            tab_index: model.tab_index as usize,
         }
     }
 }
